% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_function.R
\name{cv.cluster.lm}
\alias{cv.cluster.lm}
\title{Cluster Function}
\usage{
cv.cluster.lm(Data, formulae, nfolds = 5, clusterID, N, method = "linear")
}
\description{
This is a cross validation function designed for samples taken using a
clustered sampling design
}
