% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs_function.R
\name{cv.srs.lm}
\alias{cv.srs.lm}
\title{SRS Function}
\usage{
cv.srs.lm(Data, formulae, nfolds = 5, N, method = "linear")
}
\description{
This is a cross validation function for survey samples taken using a a simple
random sample survey design
}
