% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_function.R
\name{cv.strat.lm}
\alias{cv.strat.lm}
\title{Stratified function}
\usage{
cv.strat.lm(Data, formulae, nfolds = 5, strataID, N, method = "linear")
}
\description{
this is a cross validation function designed for survey samples taken using a
stratified sampling design
}
