% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.svy.R
\name{cv.svy}
\alias{cv.svy}
\title{CV for survey data}
\usage{
cv.svy(
  Data,
  formulae,
  nfolds = 5,
  strataID = NULL,
  clusterID = NULL,
  nest = FALSE,
  fpcID = NULL,
  method = c("linear", "logistic"),
  weightsID = NULL,
  useSvyForFolds = TRUE,
  useSvyForFits = TRUE,
  useSvyForLoss = TRUE
)
}
\arguments{
\item{Data}{Dataframe of dataset to be tested}

\item{formulae}{Vector of formulas (as strings) for the GLMs to be compared in
cross validation}

\item{nfolds}{Number of folds to be used during cross validation, defaults to
5}

\item{strataID}{String of the variable name used to stratify during sampling, must
be the same as in the dataset used}

\item{clusterID}{String of the variable name used to cluster during sampling, must
be the same as in the dataset used}

\item{nest}{Specify nest = TRUE if clusters are nested within strata, defaults to FALSE}

\item{fpcID}{String of the variable name used for finite population corrections, must
be the same as in the dataset used, see `?svydesign` for details}

\item{method}{String, must be either "linear" or "logistic", determines type of
model fit during cross validation, defaults to linear}

\item{weightsID}{String of the variable name in the dataset that contains sampling weights}

\item{useSvyForFolds}{Specify useSvyForFolds = TRUE (default) to take svydesign into account when making folds;
should not be set FALSE except for running simulations to understand the properties of surveyCV}

\item{useSvyForFits}{Specify useSvyForFits = TRUE (default) to take svydesign into account when fitting models on training sets;
should not be set FALSE except for running simulations to understand the properties of surveyCV}

\item{useSvyForLoss}{Specify useSvyForLoss = TRUE (default) to take svydesign into account when calculating loss over test sets;
should not be set FALSE except for running simulations to understand the properties of surveyCV}
}
\description{
This is a cross validation function designed for survey samples taken using a SRS,
stratified, clustered, or clustered-and-stratified sampling design.
}
\examples{
# Compare CV MSEs and their SEs under 3 models
# for a stratified sample and a one-stage cluster sample,
# from the `survey` package
data("api", package = "survey")
# stratified sample
cv.svy(apistrat, c("api00~ell",
                   "api00~ell+meals",
                   "api00~ell+meals+mobility"),
       nfolds = 5, strataID = "stype", weightsID = "pw", fpcID = "fpc")
# one-stage cluster sample
cv.svy(apiclus1, c("api00~ell",
                   "api00~ell+meals",
                   "api00~ell+meals+mobility"),
       nfolds = 5, clusterID = "dnum", weightsID = "pw", fpcID = "fpc")

# Compare CV MSEs and their SEs under 3 models
# for a stratified cluster sample with clusters nested within strata
data(NSFG_data)
library(splines)
cv.svy(NSFG_data, c("income ~ ns(age, df = 2)",
                    "income ~ ns(age, df = 3)",
                    "income ~ ns(age, df = 4)"),
       nfolds = 4,
       strataID = "strata", clusterID = "SECU",
       nest = TRUE, weightsID = "wgt")
}
