% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.svydesign.R
\name{cv.svydesign}
\alias{cv.svydesign}
\title{CV for `svydesign` objects}
\usage{
cv.svydesign(
  design_object,
  formulae,
  nfolds = 5,
  method = c("linear", "logistic")
)
}
\arguments{
\item{design_object}{Name of a `svydesign` object created using the `survey`
package. The argument `ids` (also `strata`, `fpc`, and/or `weights` if used)
must be specified as formulas, e.g. `svydesign(ids = ~MyPSUs, ...)`.
We do not yet support use of `probs` or `pps`.}

\item{formulae}{Vector of formulas (as strings) for the GLMs to be compared in
cross validation}

\item{nfolds}{Number of folds to be used during cross validation, defaults to
5}

\item{method}{String, must be either "linear" or "logistic", determines type of
model fit during cross validation, defaults to linear}
}
\description{
Wrapper function which takes a `svydesign` object
and a vector of model formulas (as strings),
and passes it into `cv.svy()`.
}
\examples{
# MSEs generated for different tests of first and second degree polynomial
# fits predicting mpg from horsepower in the Auto dataset. Clustering and
# Stratification was done along the "year" variable.
# Using a survey design object to generate MSEs.
data("Auto", package = "ISLR")
library(survey)
auto.srs.svy <- svydesign(ids = ~0,
                         data = Auto)
auto.clus.svy <- svydesign(ids = ~year,
                          data = Auto)
auto.strat.svy <- svydesign(ids = ~0,
                           strata = ~year,
                           data = Auto)
cv.svydesign(formulae = c("mpg~poly(horsepower,1, raw = TRUE)",
                          "mpg~poly(horsepower,2, raw = TRUE)",
                          "mpg~poly(horsepower,3, raw = TRUE)"),
             design_object = auto.srs.svy, nfolds = 10)
cv.svydesign(formulae = c("mpg~poly(horsepower,1, raw = TRUE)",
                          "mpg~poly(horsepower,2, raw = TRUE)",
                          "mpg~poly(horsepower,3, raw = TRUE)"),
             design_object = auto.clus.svy, nfolds = 10)
cv.svydesign(formulae = c("mpg~poly(horsepower,1, raw = TRUE)",
                          "mpg~poly(horsepower,2, raw = TRUE)",
                          "mpg~poly(horsepower,3, raw = TRUE)"),
             design_object = auto.strat.svy, nfolds = 10)
}
