% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NSFG_data_everypreg}
\alias{NSFG_data_everypreg}
\title{Subset of the 2015-2017 National Survey of Family Growth (NSFG): all live births per respondent.}
\format{
A data frame with 5089 rows and 17 variables
}
\usage{
NSFG_data_everypreg
}
\description{
Same as `NSFG_data` but using *every* birth, not just the *first* birth,
out of the initial subset there
(live births with gestational age < 45 weeks
 for mothers aged 20 to 40 at time of conception).
}
\keyword{datasets}
