% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StatPaperSims}
\alias{StatPaperSims}
\alias{AllW}
\alias{ModW}
\alias{MSEW}
\alias{NoW}
\alias{stratsrsds}
\alias{stratstratds}
\alias{stratpopds}
\alias{clussrsds}
\alias{clusclusds}
\alias{cluspopds}
\alias{srssrsds}
\alias{srspopds}
\title{Internal datasets for plots-for-Stat-paper vignette.}
\format{
Data frames containing simulations from plots-for-Stat-paper.

An object of class \code{data.frame} with 3000 rows and 3 columns.

An object of class \code{data.frame} with 3000 rows and 3 columns.

An object of class \code{data.frame} with 3000 rows and 3 columns.

An object of class \code{data.frame} with 3000 rows and 3 columns.

An object of class \code{data.frame} with 3000 rows and 2 columns.

An object of class \code{data.frame} with 3000 rows and 2 columns.

An object of class \code{data.frame} with 3000 rows and 2 columns.

An object of class \code{data.frame} with 3000 rows and 2 columns.

An object of class \code{data.frame} with 3000 rows and 2 columns.

An object of class \code{data.frame} with 3000 rows and 2 columns.

An object of class \code{data.frame} with 3000 rows and 2 columns.

An object of class \code{data.frame} with 3000 rows and 2 columns.
}
\usage{
AllW

ModW

MSEW

NoW

stratsrsds

stratstratds

stratpopds

clussrsds

clusclusds

cluspopds

srssrsds

srspopds
}
\description{
See plots-for-Stat-paper vignette for details.
}
\keyword{datasets}
