% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_function.R
\name{cv.cluster.lm}
\alias{cv.cluster.lm}
\title{Cluster Function}
\usage{
cv.cluster.lm(
  Data,
  formulae,
  nfolds = 5,
  clusterID,
  N,
  method = "linear",
  weights = NULL
)
}
\arguments{
\item{Data}{CSV of dataset to be tested}

\item{formulae}{Vector of formulas containing the variables to be compared in
cross validation}

\item{nfolds}{Number of folds to be used during cross validation, defaults to
5}

\item{clusterID}{String of the variable used to cluster during sampling, must
be the same as in the dataset used}

\item{N}{Number equal to the total population size}

\item{method}{string, must be either linear or logistic, determines type of
model fit during cross validation, defaults to linear}

\item{weights}{Variable in data set that contains PPS weights}
}
\description{
This is a cross validation function designed for samples taken using a
clustered sampling design
}
\examples{
# MSEs generated for a clustered test of a first and second degree polynomial
# fit predicting mpg from horsepower in the Auto Dataset, clustering using
# the "year" variable
data("Auto")
cv.cluster.lm(Auto, c("mpg~poly(horsepower,1,raw=TRUE)", "mpg~poly(horsepower,2,raw=TRUE)"), nfolds = 10, "year", 400)
}
