% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_function.R
\name{cv.strat.lm}
\alias{cv.strat.lm}
\title{Stratified function}
\usage{
\method{cv}{strat.lm}(Data, formulae, nfolds = 5, strataID, N, method = "linear", weights = NULL)
}
\arguments{
\item{Data}{CSV of dataset to be tested}

\item{formulae}{Vector of formulas containing the variables to be compared in
cross validation}

\item{nfolds}{Number of folds to be used during cross validation, defaults to
5}

\item{strataID}{String of the variable used to Stratify during sampling, must
be the same as in the dataset used}

\item{N}{Number equal to the total population size}

\item{method}{string, must be either linear or logistic, determines type of
model fit during cross validation, defaults to linear}

\item{weights}{Variable in data set that contains PPS weights
 @examples
#MSEs generated for a stratified test of a first and second degree polynomial
# fit predicting mpg from horsepower in the Auto Dataset, Stratified on the
# "year" variable
data("Auto")
cv.strat.lm(Auto, c("mpg~poly(horsepower,1, raw = TRUE)", "mpg~poly(horsepower,2, raw = TRUE)"), nfolds = 10, "year", 400)}
}
\description{
this is a cross validation function designed for survey samples taken using a
stratified sampling design
}
