% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_function.R
\name{cv.svy.glm}
\alias{cv.svy.glm}
\title{CV Survey function}
\usage{
cv.svy.glm(
  Data,
  formulae,
  nfolds = 5,
  N,
  clusterID,
  strataID,
  method = "linear",
  sample_type = "SRS",
  design_object = NULL,
  glm = NULL
)
}
\description{
This function combines the other cross validation functions in this package
into a single function that is able to interpret datasets and also
survey design objects and survey glms
}
