% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_function.R
\name{cv.svy.glm}
\alias{cv.svy.glm}
\title{CV Survey function}
\usage{
cv.svy.glm(
  formulae,
  nfolds = 5,
  N,
  clusterID = NULL,
  strataID = NULL,
  method = "linear",
  weights = NULL,
  design_object = NULL,
  glm = NULL
)
}
\arguments{
\item{formulae}{Vector of formulas containing the variables to be compared in
cross validation}

\item{nfolds}{Number of folds to be used during cross validation, defaults to
5}

\item{N}{Number equal to the total population size}

\item{clusterID}{String of the variable used to cluster during sampling, must
be the same as in the dataset used, ignore if clustering was not used}

\item{strataID}{String of the variable used to stratify during sampling, must
be the same as in the dataset used, ignore if stratification was not used}

\item{method}{string, must be either linear or logistic, determines type of
model fit during cross validation, defaults to linear}

\item{weights}{Variable in data set that contains PPS weights
  @examples
#MSEs generated for different tests of first and second degree polynomial
# fits predicting mpg from horsepower in the Auto Dataset. Clustering and
Stratification were done along the was done along the "year" variable
#Using either a survey design object or survey glm to generate MSEs
data("Auto")
auto.srs.svy <- svydesign(ids = ~0,
                         strata = NULL,
                         fpc = rep(1000, 392),
                         data = Auto)
srs.model <- svyglm(mpg~horsepower+I(horsepower^2)+I(horsepower^3), design = auto.srs.svy)
auto.clus.svy <- svydesign(ids = ~year,
                          strata = NULL,
                          fpc = rep(1000, 392),
                          data = Auto)
clus.model <- svyglm(mpg~horsepower+I(horsepower^2)+I(horsepower^3), design = auto.clus.svy)
auto.strat.svy <- svydesign(ids = ~0,
                           strata = ~year,
                           fpc = rep(1000, 392),
                           data = Auto)
strat.model <- svyglm(mpg~horsepower+I(horsepower^2)+I(horsepower^3), design = auto.strat.svy)
cv.svy.glm(formulae = c("mpg~poly(horsepower,1, raw = TRUE)", "mpg~poly(horsepower,2, raw = TRUE)",
                       "mpg~poly(horsepower,3, raw = TRUE)"), design_object = auto.srs.svy, nfolds = 10, N = 1000)
cv.svy.glm(formulae = c("mpg~poly(horsepower,1, raw = TRUE)", "mpg~poly(horsepower,2, raw = TRUE)",
                       "mpg~poly(horsepower,3, raw = TRUE)"), design_object = auto.clus.svy, nfolds = 10, N = 1000)
cv.svy.glm(formulae = c("mpg~poly(horsepower,1, raw = TRUE)", "mpg~poly(horsepower,2, raw = TRUE)",
                       "mpg~poly(horsepower,3, raw = TRUE)"), design_object = auto.strat.svy, nfolds = 10, N = 1000)
cv.svy.glm(glm = srs.model, nfolds = 10, N = 1000)
cv.svy.glm(glm = clus.model, nfolds = 10, N = 1000)
cv.svy.glm(glm = strat.model, nfolds = 10, N = 1000)}

\item{design_object}{Name of a Survey Design Object created using the Survey
package if one is being used, defaults to NULL}

\item{glm}{Name of a Survey GLM created from the survey package if one is
to be used, defaults to NULL}

\item{Data}{CSV of dataset to be tested}

\item{sample_type}{string, must be either SRS, Cluster, or Strat, specifies
which sampling method was used, defaults to SRS}
}
\description{
This function combines the other cross validation functions in this package
into a single function that is able to interpret datasets and also
survey design objects and survey glms
}
