% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{cv.svy}
\alias{cv.svy}
\title{General CV function}
\usage{
cv.svy(
  Data,
  formulae,
  nfolds = 5,
  strataID = NULL,
  clusterID = NULL,
  nest = FALSE,
  N,
  method = "linear",
  weights = NULL,
  afolds = TRUE
)
}
\arguments{
\item{Data}{CSV of dataset to be tested}

\item{formulae}{Vector of formulas containing the variables to be compared in
cross validation}

\item{nfolds}{Number of folds to be used during cross validation, defaults to
5}

\item{strataID}{String of the variable used to Stratify during sampling, must
be the same as in the dataset used}

\item{clusterID}{String of the variable used to cluster during sampling, must
be the same as in the dataset used}

\item{nest}{Specify if nest = TRUE if clusters are nested within strata}

\item{N}{Number equal to the total population size}

\item{method}{string, must be either linear or logistic, determines type of
model fit during cross validation, defaults to linear}

\item{weights}{Variable in data set that contains PPS weights}

\item{afolds}{specify if we take into account svydesign when making folds
 @examples
#MSEs generated for a stratified test of a first and second degree polynomial
# fit predicting mpg from horsepower in the Auto Dataset, Stratified on the
# "year" variable
data("Auto")
cv.svy(Auto, c("mpg~poly(horsepower,1, raw = TRUE)", "mpg~poly(horsepower,2, raw = TRUE)"), nfolds = 10, strataID = "year", N = 400)}
}
\description{
this is a cross validation function designed for survey samples taken using a SRS,
stratified, Clustered, or Clustered and Stratified sampling design
}
