% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{cv.svy}
\alias{cv.svy}
\title{General CV function}
\usage{
cv.svy(
  Data,
  formulae,
  nfolds = 5,
  strataID = NULL,
  clusterID = NULL,
  nest = FALSE,
  N,
  method = c("linear", "logistic"),
  weights = NULL,
  useSvyForFolds = TRUE
)
}
\arguments{
\item{Data}{Dataframe of dataset to be tested}

\item{formulae}{Vector of formulas for the GLMs to be compared in
cross validation}

\item{nfolds}{Number of folds to be used during cross validation, defaults to
5}

\item{strataID}{String of the variable name used to stratify during sampling, must
be the same as in the dataset used}

\item{clusterID}{String of the variable name used to cluster during sampling, must
be the same as in the dataset used}

\item{nest}{Specify nest = TRUE if clusters are nested within strata, defaults to FALSE}

\item{N}{Total finite population size}

\item{method}{String, must be either "linear" or "logistic", determines type of
model fit during cross validation, defaults to linear}

\item{weights}{String of the variable name in the dataset that contains sampling weights}

\item{useSvyForFolds}{Specify useSvyForFolds = TRUE (default) to take svydesign into account when making folds;
should not be set FALSE except for running simulations to understand the properties of surveyCV}
}
\description{
This is a cross validation function designed for survey samples taken using a SRS,
stratified, clustered, or clustered-and-stratified sampling design.
}
\examples{
# MSEs generated for a stratified test of a first and second degree polynomial
# fit predicting mpg from horsepower in the Auto dataset, stratified on the
# "year" variable
data("Auto", package = "ISLR")
cv.svy(Auto, c("mpg~poly(horsepower,1, raw = TRUE)",
               "mpg~poly(horsepower,2, raw = TRUE)"),
       nfolds = 10, strataID = "year", N = 400)
}
