% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svydesign_function.R
\name{cv.svydesign}
\alias{cv.svydesign}
\title{CV Survey Design function}
\usage{
\method{cv}{svydesign}(
  formulae,
  nfolds = 5,
  N,
  clusterID = NULL,
  strataID = NULL,
  method = "linear",
  weights = NULL,
  design_object = NULL
)
}
\arguments{
\item{formulae}{Vector of formulas containing the variables to be compared in
cross validation}

\item{nfolds}{Number of folds to be used during cross validation, defaults to
5}

\item{N}{Number equal to the total population size}

\item{clusterID}{String of the variable used to cluster during sampling, must
be the same as in the dataset used, ignore if clustering was not used}

\item{strataID}{String of the variable used to stratify during sampling, must
be the same as in the dataset used, ignore if stratification was not used}

\item{method}{string, must be either linear or logistic, determines type of
model fit during cross validation, defaults to linear}

\item{weights}{Variable in data set that contains PPS weights
  @examples
#MSEs generated for different tests of first and second degree polynomial
# fits predicting mpg from horsepower in the Auto Dataset. Clustering and
Stratification was done along the "year" variable.
#Using a survey design object to generate MSEs.
data("Auto")
auto.srs.svy <- svydesign(ids = ~0,
                         strata = NULL,
                         fpc = rep(1000, 392),
                         data = Auto)
auto.clus.svy <- svydesign(ids = ~year,
                          strata = NULL,
                          fpc = rep(1000, 392),
                          data = Auto)
auto.strat.svy <- svydesign(ids = ~0,
                           strata = ~year,
                           fpc = rep(1000, 392),
                           data = Auto)
cv.svy.glm(formulae = c("mpg~poly(horsepower,1, raw = TRUE)", "mpg~poly(horsepower,2, raw = TRUE)",
                       "mpg~poly(horsepower,3, raw = TRUE)"), design_object = auto.srs.svy, nfolds = 10, N = 1000)
cv.svy.glm(formulae = c("mpg~poly(horsepower,1, raw = TRUE)", "mpg~poly(horsepower,2, raw = TRUE)",
                       "mpg~poly(horsepower,3, raw = TRUE)"), design_object = auto.clus.svy, nfolds = 10, N = 1000)
cv.svy.glm(formulae = c("mpg~poly(horsepower,1, raw = TRUE)", "mpg~poly(horsepower,2, raw = TRUE)",
                       "mpg~poly(horsepower,3, raw = TRUE)"), design_object = auto.strat.svy, nfolds = 10, N = 1000)}

\item{design_object}{Name of a Survey Design Object created using the Survey
package if one is being used, defaults to NULL}
}
\description{
This function combines the other cross validation functions in this package
into a single function that is able to interpret datasets and also
survey design objects and survey glms
}
