% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_function.R
\name{cv.svyglm}
\alias{cv.svyglm}
\title{CV Survey GLM function}
\usage{
\method{cv}{svyglm}(
  nfolds = 5,
  N,
  clusterID = NULL,
  strataID = NULL,
  method = "linear",
  weights = NULL,
  glm = NULL
)
}
\arguments{
\item{nfolds}{Number of folds to be used during cross validation, defaults to
5}

\item{N}{Number equal to the total population size}

\item{clusterID}{String of the variable used to cluster during sampling, must
be the same as in the dataset used, ignore if clustering was not used}

\item{strataID}{String of the variable used to stratify during sampling, must
be the same as in the dataset used, ignore if stratification was not used}

\item{method}{string, must be either linear or logistic, determines type of
model fit during cross validation, defaults to linear}

\item{weights}{Variable in data set that contains PPS weights
  @examples
#MSEs generated for different tests of first and second degree polynomial
# fits predicting mpg from horsepower in the Auto Dataset. Clustering and
Stratification was done along the "year" variable.
#Using a survey glm to generate MSEs.
data("Auto")
auto.srs.svy <- svydesign(ids = ~0,
                         strata = NULL,
                         fpc = rep(1000, 392),
                         data = Auto)
srs.model <- svyglm(mpg~horsepower+I(horsepower^2)+I(horsepower^3), design = auto.srs.svy)
auto.clus.svy <- svydesign(ids = ~year,
                          strata = NULL,
                          fpc = rep(1000, 392),
                          data = Auto)
clus.model <- svyglm(mpg~horsepower+I(horsepower^2)+I(horsepower^3), design = auto.clus.svy)
auto.strat.svy <- svydesign(ids = ~0,
                           strata = ~year,
                           fpc = rep(1000, 392),
                           data = Auto)
strat.model <- svyglm(mpg~horsepower+I(horsepower^2)+I(horsepower^3), design = auto.strat.svy)
cv.svyglm(glm = srs.model, nfolds = 10, N = 1000)
cv.svyglm(glm = clus.model, nfolds = 10, N = 1000)
cv.svyglm(glm = strat.model, nfolds = 10, N = 1000)}

\item{glm}{Name of a Survey GLM created from the survey package if one is
to be used, defaults to NULL}
}
\description{
This function combines the other cross validation functions in this package
into a single function that is able to interpret datasets and also
survey design objects and survey glms
}
