% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyCV.R
\docType{package}
\name{surveyCV}
\alias{surveyCV}
\title{surveyCV: Cross Validation Based on Survey Design}
\description{
Functions to generate test error estimates using cross validation,
based on how a survey design is constructed
(SRS, clustering, stratification, and/or unequal sampling weights).
You can input linear and logistic regression models, along with data and a
type of survey design in order to get an output that can help you determine
which model best fits the data using K-fold cross validation.
Our paper on "K-Fold Cross-Validation for Complex Sample Surveys"
by Wieczorek, Guerin, and McMahon (2022)
\href{https://doi.org/10.1002/sta4.454}{<doi:10.1002/sta4.454>}
explains why differing how we take folds based on survey design is useful.
}
\details{
The code for this package seeks to create an alternative for the
\code{\link[boot:cv.glm]{boot::cv.glm}}
function, so that results correctly account for survey designs during
K-fold cross validation.
}
